+++ busTrackerApi/src/main/kotlin/busTrackerApi/routing/lines/LinesClient.kt
@@ -1,8 +1,10 @@
 package busTrackerApi.routing.lines
 
 import arrow.core.continuations.either
+import busTrackerApi.db.getItinerariesByItineraryCode
 import busTrackerApi.db.getItineraryByFullLineCode
 import busTrackerApi.db.getShapesByItineraryCode
+import busTrackerApi.exceptions.BusTrackerException.NotFound
 import busTrackerApi.extensions.getWrapped
 import busTrackerApi.routing.Response.ResponseFlowJson
 import busTrackerApi.routing.Response.ResponseJson
@@ -12,6 +14,7 @@ import io.ktor.http.*
 import io.ktor.http.content.*
 import io.ktor.server.application.*
 import io.ktor.server.plugins.cachingheaders.*
+import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.map
 import simpleJson.asJson
 
@@ -20,11 +23,25 @@ suspend fun Pipeline.getItineraries() = either {
     val direction = call.parameters.getWrapped("direction").bind().toIntOrNull()
         ?: shift<Nothing>(busTrackerApi.exceptions.BusTrackerException.BadRequest())
 
-    val itineraries = getItineraryByFullLineCode(lineCode, direction - 1)
+    val itineraries = getItineraryByFullLineCode(lineCode, direction)
         ?: getItinerariesResponse(lineCode).bind()
             .sortedBy { it.stops.count() }
             .firstOrNull { it.direction == direction - 1 }
-        ?: shift<Nothing>(busTrackerApi.exceptions.BusTrackerException.NotFound("Itinerary not found"))
+        ?: shift<Nothing>(NotFound("Itinerary not found"))
+
+    val itinerariesOrdered =
+        itineraries.copy(stops = itineraries.stops.distinctBy { it.fullStopCode to it.order }.sortedBy { it.order })
+
+    val json = itinerariesOrdered.let(::buildItineraryJson).asJson()
+    call.caching = CachingOptions(CacheControl.MaxAge(maxAgeSeconds = 60 * 60))
+    ResponseJson(json, HttpStatusCode.OK)
+}
+
+suspend fun Pipeline.getItinerariesByItineraryCode() = either {
+    val itineraryCode = call.parameters.getWrapped("itineraryCode").bind()
+
+    val itineraries =
+        getItinerariesByItineraryCode(itineraryCode).firstOrNull() ?: shift<Nothing>(NotFound("Itinerary not found"))
 
     val itinerariesOrdered =
         itineraries.copy(stops = itineraries.stops.distinctBy { it.fullStopCode to it.order }.sortedBy { it.order })