+++ busTrackerApi/src/main/kotlin/busTrackerApi/routing/lines/bus/BusLinesClient.kt
@@ -1,6 +1,7 @@
 package busTrackerApi.routing.lines.bus
 
 import arrow.core.continuations.either
+import busTrackerApi.db.getItinerariesByItineraryCode
 import busTrackerApi.db.getItineraryByFullLineCode
 import busTrackerApi.db.getRoute
 import busTrackerApi.exceptions.BusTrackerException.BadRequest
@@ -18,6 +19,7 @@ import io.ktor.http.*
 import io.ktor.http.content.*
 import io.ktor.server.application.*
 import io.ktor.server.plugins.cachingheaders.*
+import kotlinx.coroutines.flow.firstOrNull
 
 suspend fun Pipeline.getLocations() = either {
     val lineCode = call.parameters.getWrapped("lineCode").bind()
@@ -27,12 +29,8 @@ suspend fun Pipeline.getLocations() = either {
     val fullStopCode = if (stopCode != null) createStopCode(busCodMode, stopCode) else null
     val codMode = getCodModeFromLineCode(lineCode)
 
-    var itinerary = getItineraryByFullLineCode(lineCode, direction - 1)
-    //TODO save in db so we don't have to ask the shitty server again
-    if (itinerary == null) {
-        itinerary = getItinerariesResponse(lineCode).bind().firstOrNull { it.direction == direction - 1 }
-            ?: shift<Nothing>(NotFound())
-    }
+    val itinerary = getItineraryByFullLineCode(lineCode, direction) ?: shift<Nothing>(NotFound())
+
     val route = getRoute(lineCode).getOrNull()
     val simpleLineCode = route?.simpleLineCode ?: getSimpleLineCodeFromLineCode(lineCode)
     val routeCodMode = route?.codMode ?: busCodMode
@@ -41,6 +39,36 @@ suspend fun Pipeline.getLocations() = either {
         locations = getLocationsResponse(
             itinerary,
             lineCode,
+            codMode,
+            fullStopCode
+        ).bind().vehiclesLocation.vehicleLocation,
+        codMode = routeCodMode.toInt(),
+        lineCode = simpleLineCode,
+    )
+
+    val json = buildVehicleLocationJson(locations)
+
+    call.caching = CachingOptions(CacheControl.MaxAge(maxAgeSeconds = 10))
+    ResponseJson(json, HttpStatusCode.OK)
+}
+
+suspend fun Pipeline.getLocationsByItineraryCode() = either {
+    val itineraryCode = call.parameters.getWrapped("itineraryCode").bind()
+    val stopCode = call.request.queryParameters["stopCode"]
+
+    val itinerary = getItinerariesByItineraryCode(itineraryCode).firstOrNull() ?: shift<Nothing>(NotFound())
+
+    val fullStopCode = if (stopCode != null) createStopCode(busCodMode, stopCode) else null
+    val codMode = getCodModeFromLineCode(itinerary.fullLineCode)
+
+    val route = getRoute(itinerary.fullLineCode).getOrNull()
+    val simpleLineCode = route?.simpleLineCode ?: getSimpleLineCodeFromLineCode(itineraryCode)
+    val routeCodMode = route?.codMode ?: busCodMode
+
+    val locations = VehicleLocations(
+        locations = getLocationsResponse(
+            itinerary,
+            itinerary.fullLineCode,
             codMode,
             fullStopCode
         ).bind().vehiclesLocation.vehicleLocation,