+++ busTrackerApi/src/main/kotlin/busTrackerApi/routing/lines/emt/EmtLinesRouting.kt
@@ -2,10 +2,12 @@ package busTrackerApi.routing.lines.emt
 
 import busTrackerApi.extensions.handle
 import busTrackerApi.routing.lines.getItineraries
+import busTrackerApi.routing.lines.getItinerariesByItineraryCode
 import busTrackerApi.routing.lines.getShapes
 import io.ktor.server.routing.*
 
 fun Route.emtLinesRouting() = route("/emt") {
+    //TODO Deprecate lineCode and direction endpoint on next major release
     get("/{lineCode}/locations/{direction}") {
         handle { getLocations() }
     }
@@ -16,5 +18,13 @@ fun Route.emtLinesRouting() = route("/emt") {
 
     get("/shapes/{itineraryCode}") {
         handle { getShapes() }
+    }
+
+    get("/locations/{itineraryCode}") {
+        handle { getLocationsByItineraryCode() }
+    }
+
+    get("/itineraries/{itineraryCode}") {
+        handle { getItinerariesByItineraryCode() }
     }
 }